Move all project struct to the `projects` module
