+++ src/gitlab.rs
@@ -305,12 +305,12 @@ impl GitLab {
     /// NOTE: We can't search for "namespace/name", so we search for "name", and refine the match
     ///       on the namespace. This means the operation could be slow as multiple query to the
     ///       GitLab server might be required to find the right item.
-    pub fn get_project(&self, namespace: &str, name: &str) -> Result<::Project> {
+    pub fn get_project(&self, namespace: &str, name: &str) -> Result<::projects::Project> {
 
         // Closure to search for the item, possibly returning multiple match on multiple pages.
         let query_gitlab_closure = || self.projects().search(name.to_string());
         // Closure to find the right item in the found list on the page.
-        let iter_find_closure = |ref project: &::Project| project.namespace.name == namespace && project.name == name;
+        let iter_find_closure = |ref project: &::projects::Project| project.namespace.name == namespace && project.name == name;
 
         self.get_paginated_from_project(query_gitlab_closure, iter_find_closure)
     }
